backend/pyproject.toml
[tool.poetry]
name = "schedule-coordinator-backend"
version = "0.1.0"
description = "FastAPI backend for Schedule Coordinator Bot with LINE and Google Calendar integration"
authors = ["Your Name <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
sqlalchemy = "^2.0"
mysql-connector-python = "^8.2.0"
python-dotenv = "^1.0.0"
google-auth = "^2.25.2"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.1.0"
google-api-python-client = "^2.100.0"
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
aiohttp = "^3.9.1"
pyjwt = "^2.8.1"

[tool.poetry.dev-dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
httpx = "^0.25.1"
black = "^23.11.0"
pylint = "^3.0.2"
isort = "^5.12.0"
flake8 = "^6.1.0"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
